// src/app/model/Demande.model.ts
export interface Demande {
  id?: number;
  immat: string;
  marque: string;
  fournisseur: string;
  motifImmo: string;
  dateDebut: string;
  dateFin: string;
  client: string;
  createurDemande: string;
  dateCloture: string;
  statut: string;
  km: number;
  nChassi: string;
  statutFrn: string;
  positionContrat: string;
  ville: string;
  gsm: string;
  email: string;
  dateProbleme: string;
  traiteurdemand: string;
}